feat(mac-dock-sort): update condition to hostname

+++ mac/dock-sort/mac-dock-sort.sh
@@ -23,7 +23,7 @@ dock_item() {
   ', "$1"
 }
 
-if [ "$(whoami)" = "mikl" ]; then
+if [ "$(hostname)" = "mikl" ]; then
   defaults write com.apple.dock persistent-apps -array \
     "$(dock_item /Applications/Brave\ Browser.app)" \
     "$(dock_item /Applications/Google\ Chrome.app)" \
@@ -42,21 +42,23 @@ if [ "$(whoami)" = "mikl" ]; then
     "$(dock_item /Applications/Visual\ Studio\ Code.app)" \
     "$(dock_item /Applications/Notion.app)" \
     "$(dock_item /Applications/ChatGPT.app)"
-else
-  defaults write com.apple.dock persistent-apps -array \
-    "$(dock_item /Applications/Brave\ Browser.app)" \
-    "$(dock_item /Applications/Google\ Chrome.app)" \
-    "$(dock_item /Applications/Google\ Chrome\ Canary.app)" \
-    "$(dock_item /System/Applications/Messages.app)" \
-    "$(dock_item /Applications/Spotify.app)" \
-    "$(dock_item /System/Applications/Music.app)" \
-    "$(dock_item /System/Applications/Photos.app)" \
-    "$(dock_item /Applications/iTerm.app)" \
-    "$(dock_item /Applications/Visual\ Studio\ Code.app)" \
-    "$(dock_item /Applications/Visual\ Studio\ Code\ -\ Insiders.app)" \
-    "$(dock_item /Applications/Postman.app)" \
-    "$(dock_item /Applications/Notion.app)" \
-    "$(dock_item /Applications/ChatGPT.app)"
 fi
 
-killall Dock
+if [ "$(hostname)" = "QHQH24WCXG-mikl" ]; then
+  defaults write com.apple.dock persistent-apps -array \
+    "$(dock_item /Applications/Google\ Chrome.app)" \
+    "$(dock_item /Applications/Google\ Chrome\ Canary.app)" \
+    "$(dock_item /Applications/iTerm.app)" \
+    "$(dock_item /Applications/Visual\ Studio\ Code.app)" \
+    "$(dock_item /Applications/Visual\ Studio.app)" \
+    "$(dock_item /Applications/Docker.app)" \
+    "$(dock_item /Applications/Postman.app)" \
+    "$(dock_item /Applications/Parallels\ Desktop.app)" \
+    "$(dock_item /Applications/Figma.app)" \
+    "$(dock_item /Applications/Notion.app)" \
+    "$(dock_item /Applications/ClickUp.app)" \
+    "$(dock_item /Applications/zoom\.us.app)" \
+    "$(dock_item /Applications/Slack.app)" \
+    "$(dock_item /Applications/ChatGPT.app)"
+  killall Dock
+fi
